Divide the four oversized wow-data catalog modules (#691)
Continues #689's crate-by-crate physical pass. Each of the four mixed several
DB2 catalogs, or one large reader with its decoding helpers:

| file | lines | submodules | largest |
| --- | --- | --- | --- |
| progression_rewards.rs | 1,292 | faction, quest, reward_pack, curves, scaling, criteria, loading | 320 |
| wdc4.rs | 1,134 | reader, format, bits | 759 |
| item_stats.rs | 1,051 | store, entries | 792 |
| player_condition.rs | 1,038 | entry, store, evaluate | 511 |

No wow-data file is over 1,000 lines now, and the physical ratchet's cohesion
reviews drop from 86 to 82.

Two mechanics this crate needed that the packet families did not:

- `macro_rules!` definitions and their top-level invocations stay in the
  module root. `db2_store!` and `impl_from_entries!` declare the store types,
  so expanding them from a submodule would move those types out of the parent
  and hide their private fields from the sibling impls. Keeping the expansion
  site where the single file had it keeps every generated type and field
  exactly where it was.
- Fields and one associated function that a sibling submodule or the module's
  own test root reads become `pub(super)`, which is the reach they already
  had: private in a single file meant visible throughout that module. The
  three WDC4 header/reader structs are widened field-for-field because they
  are built with struct literals, and E0451 requires every field visible at
  the construction site - the compiler's own diagnostic drove each change.

`SectionHeader` needed widening for a leak no textual scan can see: it is
never named in reader.rs, only inferred from `parse_section_header`, so only
the "private type in public interface" error revealed it.

Item multiset identical per divided file (95, 25, 17 and 36 items). The 724
crate tests pass with no test path renamed, and no target gains a compiler
warning. Architecture check, self-test, physical ratchet, hotspot ratchet and
the syntax-only session ownership check pass.

Closes #691
