+++ crates/wow-data/src/progression_rewards/loading.rs
@@ -1,0 +1,45 @@
+//! Loading packets.
+//!
+//! Separated from progression_rewards.rs under #691.
+
+use super::*;
+
+pub(super) fn load_store<T, S>(
+    data_dir: &str,
+    locale: &str,
+    file_name: &str,
+    mut read: impl FnMut(u32, usize, &Wdc4Reader) -> T,
+) -> Result<S>
+where
+    S: FromEntries<T>,
+{
+    let path = Path::new(data_dir).join("dbc").join(locale).join(file_name);
+    let reader =
+        Wdc4Reader::open(&path).with_context(|| format!("failed to open {}", path.display()))?;
+
+    let mut entries = Vec::with_capacity(reader.total_count());
+    for (id, idx) in reader.iter_records() {
+        entries.push(read(id, idx, &reader));
+    }
+
+    let store = S::from_entries(entries);
+    info!("Loaded {} rows from {}", store.len(), path.display());
+    Ok(store)
+}
+
+pub(super) fn f32_field(reader: &Wdc4Reader, record_idx: usize, field: usize) -> f32 {
+    f32::from_bits(reader.get_field_u32(record_idx, field))
+}
+
+pub(super) fn f32_array<const N: usize>(
+    reader: &Wdc4Reader,
+    record_idx: usize,
+    field: usize,
+) -> [f32; N] {
+    std::array::from_fn(|i| f32::from_bits(reader.get_array_element(record_idx, field, i, 32)))
+}
+
+pub(super) trait FromEntries<T> {
+    fn from_entries(entries: impl IntoIterator<Item = T>) -> Self;
+    fn len(&self) -> usize;
+}
